test: split pure logic test suites
